Adds prop-types for all components
