--- src/components/app.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-import {Main} from "./main";
-
-export const App = (props) => {
-  return (
-    // eslint-disable-next-line react/prop-types
-    <Main quantity={props.quantity}/>
-  );
-};
